pyproject.toml
[tool.poetry]
name = "lesson21-django"
version = "0.1.0"
description = ""
authors = ["ig-rudenko <ig.rudenko1@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
pillow = "^10.2.0"
psycopg2 = "^2.9.9"
django-ckeditor = "^6.7.1"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
faker = "^23.1.0"
django-debug-toolbar = "^4.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
